+++ apps/frontend/src/app/periods/[id]/page.tsx
@@ -1,6 +1,4 @@
 'use client';
-import { pdf } from '@react-pdf/renderer';
-import { ReactElement } from 'react';
 
 import { columns } from '@/app/periods/[id]/columns';
 import { DataTable } from '@/app/periods/[id]/data-table';
@@ -11,7 +9,7 @@ import LoadingCard from '@/components/ui/LoadingCard';
 import useApplications from '@/hooks/useApplications';
 import { usePeriod } from '@/hooks/usePeriod';
 import { toast } from '@/lib/use-toast';
-import { getStatusKey } from '@/lib/utils';
+import { downloadPdf, getStatusKey } from '@/lib/utils';
 import { ApplicationEntity, ApplicationStatus } from '@/types/application-entity';
 
 import { ApplicationExport } from './application-export';
@@ -55,23 +53,6 @@ export default function Page({ params }: { params: { id: number } }) {
         `schbody_applications_export_${Date.now()}.pdf`
       );
     }
-  };
-
-  const downloadPdf = async (pdfComponent: ReactElement, fileName: string) => {
-    const blob = await pdf(pdfComponent).toBlob();
-    // eslint-disable-next-line no-undef
-    const a = document.createElement('a');
-    a.style.display = 'none';
-    // eslint-disable-next-line no-undef
-    document.body.appendChild(a);
-
-    // eslint-disable-next-line no-undef
-    const url = window.URL.createObjectURL(blob);
-    a.href = url;
-    a.download = fileName;
-    a.click();
-    // eslint-disable-next-line no-undef
-    window.URL.revokeObjectURL(url);
   };
 
   if (period?.error) return <div>Hiba történt: {period?.error.message}</div>;
