+++ apps/frontend/src/lib/utils.ts
@@ -1,5 +1,7 @@
+import { pdf } from '@react-pdf/renderer';
 import { type ClassValue, clsx } from 'clsx';
 import { decode, JwtPayload } from 'jsonwebtoken';
+import { ReactElement } from 'react';
 import { twMerge } from 'tailwind-merge';
 
 import { ApplicationEntity, ApplicationStatus } from '@/types/application-entity';
@@ -76,4 +78,21 @@ export const mockApplication: ApplicationEntity = {
     role: Role.USER,
     updatedAt: '',
   },
+};
+
+export const downloadPdf = async (pdfComponent: ReactElement, fileName: string) => {
+  const blob = await pdf(pdfComponent).toBlob();
+  // eslint-disable-next-line no-undef
+  const a = document.createElement('a');
+  a.style.display = 'none';
+  // eslint-disable-next-line no-undef
+  document.body.appendChild(a);
+
+  // eslint-disable-next-line no-undef
+  const url = window.URL.createObjectURL(blob);
+  a.href = url;
+  a.download = fileName;
+  a.click();
+  // eslint-disable-next-line no-undef
+  window.URL.revokeObjectURL(url);
 };
